add useFeltLayers hook
